add auditing features
